ref: Move OTel setup out of `integrations/opentelemetry/` (#4277)
Moving stuff out of `integrations/opentelemetry/` step by step since
there is no OpenTelemetry integration anymore -- it's part of the core
SDK.

- Moved `sentry_sdk/integrations/opentelemetry/integration.py` ->
`sentry_sdk/opentelemetry/tracing.py`.
- Removed all the experimental autoinstrumentation stuff from
`integration.py`.
- Removed `integrations/opentelemetry/` altogether (there was nothing
left but `__init__.py`, which is now also gone).
- Moved all tests from `tests/integrations/opentelemetry` to
`tests/opentelemetry`.
- Removed the potel/opentelemetry integration test targets from
`tox.ini`. These will now be run as part of the Common test suite.

Ref https://github.com/getsentry/sentry-python/issues/3853
