+++ .github/workflows/test-integrations-misc.yml
@@ -29,7 +29,7 @@ jobs:
     strategy:
       fail-fast: false
       matrix:
-        python-version: ["3.7","3.8","3.9","3.10","3.11","3.12","3.13"]
+        python-version: ["3.7","3.8","3.10","3.11","3.12","3.13"]
         os: [ubuntu-22.04]
     steps:
       - uses: actions/checkout@v4.2.2
