+++ scripts/populate_tox/tox.jinja
@@ -107,12 +107,6 @@ envlist =
     {py3.9,py3.11,py3.12}-openai-latest
     {py3.9,py3.11,py3.12}-openai-notiktoken
 
-    # OpenTelemetry (OTel)
-    {py3.7,py3.9,py3.12,py3.13}-opentelemetry
-
-    # OpenTelemetry Experimental (POTel)
-    {py3.8,py3.9,py3.10,py3.11,py3.12,py3.13}-potel
-
     # pure_eval
     {py3.7,py3.12,py3.13}-pure_eval
 
@@ -310,12 +304,6 @@ deps =
     openai-latest: tiktoken~=0.6.0
     openai-notiktoken: openai
 
-    # OpenTelemetry (OTel)
-    opentelemetry: opentelemetry-distro
-
-    # OpenTelemetry Experimental (POTel)
-    potel: -e .[opentelemetry-experimental]
-
     # pure_eval
     pure_eval: pure_eval
 
@@ -441,8 +429,6 @@ setenv =
     loguru: TESTPATH=tests/integrations/loguru
     openai: TESTPATH=tests/integrations/openai
     openfeature: TESTPATH=tests/integrations/openfeature
-    opentelemetry: TESTPATH=tests/integrations/opentelemetry
-    potel: TESTPATH=tests/integrations/opentelemetry
     pure_eval: TESTPATH=tests/integrations/pure_eval
     pymongo: TESTPATH=tests/integrations/pymongo
     pyramid: TESTPATH=tests/integrations/pyramid
