+++ tests/opentelemetry/test_entry_points.py
