+++ tox.ini
@@ -10,7 +10,7 @@
 # The file (and all resulting CI YAMLs) then need to be regenerated via
 # "scripts/generate-test-files.sh".
 #
-# Last generated: 2025-04-10T12:24:07.555804+00:00
+# Last generated: 2025-04-10T13:32:09.968256+00:00
 
 [tox]
 requires =
@@ -106,12 +106,6 @@ envlist =
     {py3.9,py3.11,py3.12}-openai-v1.55
     {py3.9,py3.11,py3.12}-openai-latest
     {py3.9,py3.11,py3.12}-openai-notiktoken
-
-    # OpenTelemetry (OTel)
-    {py3.7,py3.9,py3.12,py3.13}-opentelemetry
-
-    # OpenTelemetry Experimental (POTel)
-    {py3.8,py3.9,py3.10,py3.11,py3.12,py3.13}-potel
 
     # pure_eval
     {py3.7,py3.12,py3.13}-pure_eval
@@ -435,12 +429,6 @@ deps =
     openai-latest: tiktoken~=0.6.0
     openai-notiktoken: openai
 
-    # OpenTelemetry (OTel)
-    opentelemetry: opentelemetry-distro
-
-    # OpenTelemetry Experimental (POTel)
-    potel: -e .[opentelemetry-experimental]
-
     # pure_eval
     pure_eval: pure_eval
 
@@ -757,8 +745,6 @@ setenv =
     loguru: TESTPATH=tests/integrations/loguru
     openai: TESTPATH=tests/integrations/openai
     openfeature: TESTPATH=tests/integrations/openfeature
-    opentelemetry: TESTPATH=tests/integrations/opentelemetry
-    potel: TESTPATH=tests/integrations/opentelemetry
     pure_eval: TESTPATH=tests/integrations/pure_eval
     pymongo: TESTPATH=tests/integrations/pymongo
     pyramid: TESTPATH=tests/integrations/pyramid
